requirements.txt
hikari
requests
beautifulsoup4
cmdtools-py
